Ever-so-slightly improve compatibility with MSVC.
Update xdr.c

+++ newlib/libc/xdr/xdr.c
@@ -781,7 +781,7 @@ xdr_string (XDR * xdrs,
       if (sp == NULL)
         return FALSE;
 
-      size = strlen (sp);
+      size = (u_int)strlen (sp);
       break;
     case XDR_DECODE:
       break;
